requirements.txt
streamlit
Pillow
pytesseract
google-generativeai
re